+++ src/main/java/com/gdsc/homework/HW5/controller/user/dto/response/UserResponse.java
@@ -1,4 +1,4 @@
-package com.gdsc.homework.HW5.service.dto.request;
+package com.gdsc.homework.HW5.controller.user.dto.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,12 +8,13 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class UserServiceRequest {
+public class UserResponse {
+    private Long id;
     private String name;
     private String email;
     private String password;
 
-    public static UserServiceRequest newInstance(String name, String email, String password){
-        return new UserServiceRequest(name,email,password);
+    public static UserResponse newInstance(Long id, String name, String email, String password) {
+        return new UserResponse(id, name, email, password);
     }
 }
