+++ src/main/java/com/gdsc/homework/HW5/service/user/dto/request/UserServiceRequest.java
@@ -1,4 +1,4 @@
-package com.gdsc.homework.HW5.controller.dto.response;
+package com.gdsc.homework.HW5.service.user.dto.request;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,13 +8,12 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class UserResponse {
-    private Long id;
+public class UserServiceRequest {
     private String name;
     private String email;
     private String password;
 
-    public static UserResponse newInstance(Long id, String name, String email, String password) {
-        return new UserResponse(id, name, email, password);
+    public static UserServiceRequest newInstance(String name, String email, String password){
+        return new UserServiceRequest(name,email,password);
     }
 }
